requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.4.0
cssselect2==0.7.0
Django==3.2.25
django-environ==0.11.2
django-ninja==1.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
dnspython==2.3.0
email-validator==1.1.1
fonttools==4.38.0
html5lib==1.1
idna==3.10
importlib-metadata==6.7.0
Pillow==9.5.0
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
pydyf==0.9.0
PyJWT==2.8.0
pypdf2==3.0.1
pyphen==0.14.0
pytz==2024.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing-extensions==4.7.1
urllib3==2.0.7
weasyprint==60.2
webencodings==0.5.1
zipfile36==0.1.3
zipp==3.15.0
zopfli==0.2.2
